requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-image>=0.22.0
Pillow>=10.0.0
tqdm>=4.66.0
joblib>=1.3.0
pyyaml>=6.0.0
streamlit>=1.38.0
plotly>=5.18.0